Added NP_YUI Minor fix in NPLib_Security

+++ php/NPLib_YUI.php
@@ -1,21 +1,64 @@
 <?
-function NP_YUI_generateJS($class, $extraOptions = null) {
-   $lower_class = strtolower($class);
+/**
+ *
+ * NPLib_YUI available event hooks:
+ * {DATA event}
+ *  - Insert: 
+ *       boolean beforeInsert(formObject)
+ *       boolean afterInsert()
+ *  - Delete: 
+ *       boolean beforeDelete(list)
+ *       boolean afterDelete()
+ *  - Update: 
+ *       boolean beforeUpdate(record, column, oldValue, newValue)  
+ *       boolean afterUpdate()
+ *  - Load: 
+ *       void onLoad(response), 
+ *
+ * {UI event}
+ *  - InsertDialog: 
+ *       boolean beforeInsertDialog()
+ *       boolean afterInsertDialog(formObject)
+ *
+ */
+
+class NP_YUI {
    
-   if ($extraOptions == null) 
-       $extraOptions = array();
-       
-   if (!array_key_exists("ajaxURL", $extraOptions))
-       $extraOptions["ajaxURL"] = "./ajax/".$lower_class.".php";
-   if (!array_key_exists("idFields", $extraOptions))
-       $extraOptions["idFields"] = "id";
+   public function __construct($className, $options = null) {
+      $lower_class = strtolower($className);
+      
+      if (!array_key_exists("ajaxURL", $options))
+          $options["ajaxURL"] = "./ajax/".$lower_class.".php";
+      if (!array_key_exists("idFields", $options))
+          $options["idFields"] = "id";
+          
+      $this->className = $className;
+      $this->options = $options;
+
+   }
+   
+   public function setColumns($columns) {
+      $this->columns = $columns;
+   }
+   
+   public function setDataFields($fields) {
+      $this->fields = $fields;
+   }
+   
+   public function generateJS() {
+      $class = $this->className;
+      $lower_class = strtolower($class);
+      if ($this->options == null)
+         $extraOptions = array();
+      else 
+         $extraOptions = $this->options;
 ?>
    var <?= $lower_class ?>_dataSource = null;
    var <?= $lower_class ?>_datatable = null;
    var <?= $class ?>AddDialog = null;
    
    function init_<?= $class ?>(<?= $class ?>ColumnDefs, <?= $class ?>Fields, <?= $class ?>Hooks) {
-      	 
+          
       <?= $lower_class ?>_dataSource = new YAHOO.util.DataSource("<?= $extraOptions["ajaxURL"] ?>" + "?");
       <?= $lower_class ?>_dataSource.connMethodPost = true;
       <?= $lower_class ?>_dataSource.responseType = YAHOO.util.DataSource.TYPE_JSON; 
@@ -28,8 +71,8 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
          var count = <?= $lower_class ?>_datatable.getRecordSet().getLength();
          <?= $lower_class ?>_datatable.deleteRows(0,count);
          
-         if (<?= $class ?>Hooks["onLoadingData"] != null)
-            <?= $class ?>Hooks["onLoadingData"](oParsedResponse.results);
+         if (<?= $class ?>Hooks["onLoad"] != null)
+            <?= $class ?>Hooks["onLoad"](oParsedResponse);
          
          return oParsedResponse;
       }
@@ -69,7 +112,7 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
    }
    
    function show<?= $class ?>AddDialog() {
-      if (<?= $class ?>Hooks["checkCanCreate"] == null || <?= $class ?>Hooks["checkCanCreate"] != null && <?= $class ?>Hooks["checkCanCreate"]()) {
+      if (<?= $class ?>Hooks["beforeInsertDialog"] == null || <?= $class ?>Hooks["beforeInsertDialog"] != null && <?= $class ?>Hooks["beforeInsertDialog"]()) {
          <?= $class ?>AddDialog.render(document.body);
          <?= $class ?>AddDialog.show();
       }
@@ -77,10 +120,8 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
    
    function add<?= $class ?>() {
       var formObject = document.getElementById('<?= $class ?>_form');
-      if (formObject.name.value.trim().length == 0)
-         box_block("<?= $class ?>add_block", "All the required fields have to be filled");
-      else {         
-         if (<?= $class ?>Hooks["checkDataOnCreate"] == null || <?= $class ?>Hooks["checkDataOnCreate"] != null && <?= $class ?>Hooks["checkDataOnCreate"](formObject)) {
+      if (<?= $class ?>Hooks["afterInsertDialog"] == null || <?= $class ?>Hooks["afterInsertDialog"] != null && <?= $class ?>Hooks["afterInsertDialog"](formObject)) {
+         if (<?= $class ?>Hooks["beforeInsert"] == null || <?= $class ?>Hooks["beforeInsert"] != null && <?= $class ?>Hooks["beforeInsert"](formObject)) {
             YAHOO.util.Connect.setForm(formObject);    
             var transaction = YAHOO.util.Connect.asyncRequest('POST', "<?= $extraOptions["ajaxURL"] ?>", {success:add<?= $class ?>Callback});
          }
@@ -93,11 +134,13 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
          <?= $class ?>AddDialog.hide();   
          var formObject = document.getElementById('<?= $class ?>_form');
          formObject.reset();
-   
-         var count = <?= $lower_class ?>_datatable.getRecordSet().getLength();
-         <?= $lower_class ?>_datatable.deleteRows(0,count);
-   
-         <?= $lower_class ?>_dataSource.sendRequest("op=list", {success : <?= $lower_class ?>_datatable.onDataReturnAppendRows, scope: <?= $lower_class ?>_datatable})
+         
+         if (<?= $class ?>Hooks["afterInsert"] == null || <?= $class ?>Hooks["afterInsert"] != null && <?= $class ?>Hooks["afterInsert"]()) {
+            var count = <?= $lower_class ?>_datatable.getRecordSet().getLength();
+            <?= $lower_class ?>_datatable.deleteRows(0,count);
+      
+            <?= $lower_class ?>_dataSource.sendRequest("op=list", {success : <?= $lower_class ?>_datatable.onDataReturnAppendRows, scope: <?= $lower_class ?>_datatable})
+         }
    
       } else {
          box_error("<?= $class ?>add_result", response.responseText);
@@ -119,39 +162,42 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
          for (var id in rows) {  
             var record = <?= $lower_class ?>_datatable.getRecord(rows[id]);
             if (record != null) {
-         <?
-         $fields = split(",", $extraOptions["idFields"]);
-         foreach ($fields as $f) {
-         	if ($f != null && $f != "") {
-         ?>            
-                list += record.getData("<?= trim($f) ?>");
-         <?
-         		if (count($fields) > 0) {
-         ?>
-         	    list += "#";
-         <?
-         		}
-         	}
-         }
-         ?>
-         		list = list.substring(0, list.length - 1) + ",";               
+            <?
+            $fields = split(",", $extraOptions["idFields"]);
+            foreach ($fields as $f) {
+               if ($f != null && $f != "") {
+            ?>            
+               list += record.getData("<?= trim($f) ?>");
+            <?
+            if (count($fields) > 0) {
+            ?>
+               list += "#";
+            <?
+                  }
+               }
+            }
+            ?>
+               list = list.substring(0, list.length - 1) + ",";               
             }
          }
          
          list = list.substring(0, list.length - 1);
          this.hide();
       }
-     
-      var postdata = "op=delete&list=" + list;
-      var transaction = YAHOO.util.Connect.asyncRequest('POST', "<?= $extraOptions["ajaxURL"] ?>", {success:delete<?= $class ?>Callback}, postdata);
+      if (<?= $class ?>Hooks["beforeDelete"] == null || <?= $class ?>Hooks["beforeDelete"] != null && <?= $class ?>Hooks["beforeDelete"](list)) {
+         var postdata = "op=delete&list=" + list;
+         var transaction = YAHOO.util.Connect.asyncRequest('POST', "<?= $extraOptions["ajaxURL"] ?>", {success:delete<?= $class ?>Callback}, postdata);
+      }
    }
    
    function delete<?= $class ?>Callback(response) {
       if (response.responseText.trim() == "OK") {
-         var count = <?= $lower_class ?>_datatable.getRecordSet().getLength();
-         <?= $lower_class ?>_datatable.deleteRows(0, count);
-   
-         <?= $lower_class ?>_dataSource.sendRequest("op=list", {success : <?= $lower_class ?>_datatable.onDataReturnAppendRows, scope: <?= $lower_class ?>_datatable})
+         if (<?= $class ?>Hooks["afterDelete"] == null || <?= $class ?>Hooks["afterDelete"] != null && <?= $class ?>Hooks["afterDelete"]()) {
+            var count = <?= $lower_class ?>_datatable.getRecordSet().getLength();
+            <?= $lower_class ?>_datatable.deleteRows(0, count);
+      
+            <?= $lower_class ?>_dataSource.sendRequest("op=list", {success : <?= $lower_class ?>_datatable.onDataReturnAppendRows, scope: <?= $lower_class ?>_datatable})
+         }
       } else {
          box_error("<?= $class ?>del_result", response.responseText);
       }
@@ -164,36 +210,100 @@ function NP_YUI_generateJS($class, $extraOptions = null) {
       var datatable = this.getDataTable();
       
       if (oldValue != newValue) {
-         var postdata = "op=update";
-         <?
-         $fields = split(",", $extraOptions["idFields"]);
-         foreach ($fields as $f) {
-         	if ($f != null && $f != "") {
-         ?>
-         postdata += "&<?= trim($f) ?>=" + record.getData("<?= trim($f) ?>");
-         <?
-         	}
+         if (<?= $class ?>Hooks["beforeUpdate"] == null || <?= $class ?>Hooks["beforeUpdate"] != null && <?= $class ?>Hooks["beforeUpdate"](record, column, oldValue, newValue)) {
+            var postdata = "op=update";
+            <?
+            $fields = split(",", $extraOptions["idFields"]);
+            foreach ($fields as $f) {
+               if ($f != null && $f != "") {
+            ?>
+            postdata += "&<?= trim($f) ?>=" + record.getData("<?= trim($f) ?>");
+            <?
+               }
+            }
+            ?>
+            postdata += "&" + column.key + "=" + newValue;
+            
+            YAHOO.util.Connect.asyncRequest("POST", "<?= $extraOptions["ajaxURL"] ?>", {
+               success: function (o) {
+                  if (<?= $class ?>Hooks["afterUpdate"] == null || <?= $class ?>Hooks["afterUpdate"] != null && <?= $class ?>Hooks["afterUpdate"]()) {
+                     callback(true, newValue);
+                  }
+               },
+               failure: function (o) {
+                  box_error("<?= $class ?>field_update", o.responseText);
+                  callback();
+               },
+               scope: this
+            }, postdata);
          }
-         ?>
-         postdata += "&" + column.key + "=" + newValue;
-         
-         YAHOO.util.Connect.asyncRequest("POST", "<?= $extraOptions["ajaxURL"] ?>", {
-            success: function (o) {
-               callback(true, newValue);
-            },
-            failure: function (o) {
-               box_error("<?= $class ?>field_update", o.responseText);
-               callback();
-            },
-            scope: this
-         }, postdata);
       } else {
          callback(true, oldValue);
       }
    }
+<?       
+   }
+   
+   public function generateInitJS($hooks = null) {
+      $class = $this->className;
+      $lower_class = strtolower($class);
+      
+      $columnsJS = "";
+   	  foreach ($this->columns as $name => $data) {
+		 $columnsJS .= '{ key: "'.$name.'", ';
+		 if (array_key_exists("label", $data)) 
+		 	$columnsJS .= 'label: "'.$data['label'].'", ';
+		 if (array_key_exists("sortable", $data)) 
+		 	$columnsJS .= 'sortable: '.($data['sortable'] ? "true" : "false").', ';		 	
+		 if (array_key_exists("editor", $data)) 
+		 	$columnsJS .= 'editor: '.$data['editor'].', ';
+		 $columnsJS .= "}, ";		 	
+	  }
+	  
+	  $fieldsJS = "";
+	  foreach ($this->fields as $f) {
+	  	 $fieldsJS .= '"'.$f.'", ';
+	  }
+?>
+      <?= $class ?>ColumnDefs = [ <?= $columnsJS ?> ];
 
-<? 
+      <?= $class ?>Fields = [ <?= $fieldsJS ?> ];
+      
+      init_<?= $class ?>(<?= $class ?>ColumnDefs,<?= $class ?>Fields, <?= $hooks ?>);
+<?      
+   }
+   
+   public function generateCSS() {
+   	  $class = $this->className;
+      $lower_class = strtolower($class);
+?>
 
+   #<?= $class ?>_buttons {
+       border: 1px solid black;
+       margin-bottom:10px;
+       padding:10px;
+       background-color: rgb(190,211,206);
+   }
+
+   #<?= $class ?>_datatable {
+      margin-top: 10px;
+      margin-bottom: 10px;
+   }
+   
+   #<?= $class ?>_form_table td {
+      padding: 3px;
+   }
+
+   .yui-button#<?= $class ?>DelButton button {
+      padding-left: 2em;
+      background: url(<?= npadmin_setting('NP-ADMIN', 'BASE_URL') ?>/static/img/del.gif) 5% 50% no-repeat;
+   }
+   .yui-button#<?= $class ?>AddButton button {
+      padding-left: 2em;
+      background: url(<?= npadmin_setting('NP-ADMIN', 'BASE_URL') ?>/static/img/add.gif) 5% 50% no-repeat;
+   }
+<?      
+   }
 }
 
 function NP_YUI_generateHTML($class) {
@@ -213,33 +323,6 @@ function NP_YUI_generateHTML($class) {
   </div>
 </div>
 
-<?
-}
-
-function NP_YUI_generateCSS($class) {
-	$lower_class = strtolower($class);
-?>
-
-   #<?= $class ?>_buttons {
-	    border: 1px solid black;
-	    margin-bottom:10px;
-	    padding:10px;
-	    background-color: rgb(190,211,206);
-   }
-
-   #<?= $lower_class ?>_datatable {
-	   margin-top: 10px;
-	   margin-bottom: 10px;
-   }
-
-   .yui-button#<?= $class ?>DelButton button {
-      padding-left: 2em;
-      background: url(<?= npadmin_setting('NP-ADMIN', 'BASE_URL') ?>/static/img/del.gif) 5% 50% no-repeat;
-   }
-   .yui-button#<?= $class ?>AddButton button {
-      padding-left: 2em;
-      background: url(<?= npadmin_setting('NP-ADMIN', 'BASE_URL') ?>/static/img/add.gif) 5% 50% no-repeat;
-   }
 <?
 }
 ?>
